requirements.txt
fastapi
uvicorn
sqlmodel
fastapi-users
python-jose[cryptography]
passlib[bcrypt]
asyncmy
pymysql
pandas
numpy
scikit-learn
python-dotenv
